statistics: refactor test package (#47608)
ref pingcap/tidb#44940
