+++ pkg/statistics/handle/globalstats/BUILD.bazel
@@ -34,17 +34,28 @@ go_library(
 go_test(
     name = "globalstats_test",
     timeout = "short",
-    srcs = ["topn_bench_test.go"],
-    embed = [":globalstats"],
+    srcs = [
+        "globalstats_test.go",
+        "main_test.go",
+        "topn_bench_test.go",
+    ],
     flaky = True,
+    shard_count = 14,
     deps = [
+        ":globalstats",
+        "//pkg/config",
+        "//pkg/parser/model",
         "//pkg/parser/mysql",
         "//pkg/sessionctx/stmtctx",
         "//pkg/statistics",
+        "//pkg/testkit",
+        "//pkg/testkit/testsetup",
         "//pkg/types",
         "//pkg/util/chunk",
         "//pkg/util/codec",
+        "@com_github_pingcap_failpoint//:failpoint",
         "@com_github_stretchr_testify//require",
         "@com_github_tiancaiamao_gp//:gp",
+        "@org_uber_go_goleak//:goleak",
     ],
 )
