+++ pkg/statistics/handle/globalstats/globalstats_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package globalstats
+package globalstats_test
 
 import (
 	"fmt"
