--- pkg/statistics/handle/handletest/globalstats/BUILD.bazel
@@ -1,21 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_test")
-
-go_test(
-    name = "globalstats_test",
-    timeout = "short",
-    srcs = [
-        "globalstats_test.go",
-        "main_test.go",
-    ],
-    flaky = True,
-    shard_count = 14,
-    deps = [
-        "//pkg/config",
-        "//pkg/parser/model",
-        "//pkg/testkit",
-        "//pkg/testkit/testsetup",
-        "@com_github_pingcap_failpoint//:failpoint",
-        "@com_github_stretchr_testify//require",
-        "@org_uber_go_goleak//:goleak",
-    ],
-)
