+++ pkg/statistics/handle/lockstats/main_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package globalstats
+package lockstats
 
 import (
 	"testing"
